Not allow to move along the edges of the pitch
Currently ball can be moved across every dot on the pitch. This is
unacceptable since on the starting screen player can see border of
the pitch.

This commit will prevent player to make a moves along the edges of
the pitch. Currently only left and right edge is protected. Player
can still move along the goal area and other drawn lines.